Web: Refactor build script to run in parallel with full LTO
Fabio found that full LTO produces smaller binaries than ThinLTO,
and that's quite important for the Web platform.

But for an obscure reason LLVM's full LTO cannot linking with multiple
threads, and so it's slow as heck (10-15 min per build).

So we work it around by starting all builds in parallel so that they
can all link at the same time.

The code to do so is pretty ugly, could be refactored further.